+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Breaking
+
+- removed unnecessary newtype `AccountDataRef` for `ReadonlyAccountData` impl for `solana_sdk::account::Account`
+
+### Added
+
 ## [1.1.0]
 
 ### Changed